Add collapse effect for sidebar
