+++ src/layouts/Sidebar/Sidebar.scss
@@ -1,23 +1,50 @@
 .dashboard-sidebar {
   width: 250px;
-  background-color: #f8f9fa;
-  border-right: 1px solid #e0e0e0;
-  height: calc(100vh - 120px); // Adjust based on header/footer height
+  background: rgba(255, 255, 255, 0.05);
+  border-right: 1px solid #383838;;
   display: flex;
   flex-direction: column;
   padding: 16px;
   box-sizing: border-box;
+  border-radius: 0px 24px 24px 0px;
+  transition: width 0.3s ease;
+
+  &.collapsed {
+    width: 60px;
+    padding: 16px 8px;
+  }
+
+  .toggle-sidebar-btn {
+    background-color: transparent;
+    color: white;
+    border: 1px solid rgba(255, 255, 255, 0.2);
+    border-radius: 4px;
+    padding: 6px 10px;
+    cursor: pointer;
+    font-size: 14px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 30px;
+    transition: all 0.2s ease;
+    align-self: flex-end;
+    margin-bottom: 16px;
+
+    &:hover {
+      background-color: rgba(255, 255, 255, 0.1);
+    }
+  }
 
   .sidebar-header {
     display: flex;
     justify-content: space-between;
     align-items: center;
-    margin-bottom: 16px;
+    margin: 26px 0px;
 
     h2 {
       margin: 0;
       font-size: 18px;
-      color: #333;
+      color: white;
     }
 
     .create-dashboard-btn {
@@ -94,17 +121,18 @@
       align-items: center;
       padding: 12px;
       margin-bottom: 8px;
-      background-color: #fff;
+      background: rgba(255, 255, 255, 0.05);
+      border: 1px solid #383838;
       border-radius: 4px;
       cursor: pointer;
       transition: all 0.2s ease;
 
       &:hover {
-        box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+        background: rgba(255, 255, 255, 0.1);
       }
 
       &.selected {
-        background-color: #007bff;
+        background: rgba(255, 255, 255, 0.1);
         color: white;
 
         .dashboard-name {
@@ -128,7 +156,7 @@
 
         .dashboard-name {
           font-weight: 500;
-          color: #333;
+          color: white;
         }
       }
