+++ src/pages/Dashboard/Dashboard.scss
@@ -1,7 +1,7 @@
 .board {
   min-height: 100vh;
-  background: linear-gradient(135deg, #0a0a0a 0%, #1a1a2e 100%);; // Trello blue
-  padding: 32px 0 0 0;
+  background: linear-gradient(135deg, #0a0a0a 0%, #1a1a2e 100%); // Trello blue
+  padding: 32px 0 0;
   overflow-x: auto;
 
   &-title {
@@ -24,7 +24,6 @@
   padding: 0 24px;
   min-height: 80vh;
 
-  // Responsive behavior for tablets and smartphones
   @media (max-width: 600px) {
     flex-direction: column;
     align-items: center;
@@ -44,7 +43,6 @@
     border-radius: 8px;
     padding: 7px 10px;
     font-size: 1rem;
-    // background: #fff;
     box-shadow: 0 1px 3px rgba(0,0,0,0.04);
     outline: none;
     transition: box-shadow 0.15s;
@@ -68,5 +66,14 @@
       background: rgba(255, 255, 255, 0.1);
       border-color: rgba(255, 255, 255, 0.4);
     }
+  }
+}
+
+.main-content {
+  transition: margin-left 0.3s ease;
+  margin-left: 30px;
+
+  &.sidebar-collapsed {
+    margin-left: 30px;
   }
 }
